remove 01_colors.js
